Bugfix for Wind Turbines and Solar Panels being stuck Shutdown / offline.
Bug: When the Shutdown setting is enabled as the default for buildings this causes all future constructed Wind and Solar buildings to be Shutdown without the panel option to turn back on.

Fix: A check is made prior to placing a building; if the entity is a Wind Turbine or Solar Panel then the Shutdown setting is set to off.

Result: Wind and Solar buildings cannot be Shutdown when enabling the setting as the default. If either is constructed while Shutdown is enabled as default then both will function as normal whereas all other buildings will continue to be built Shutdown / offline.