Update to InspectPage's BuildingProfit Label
-> Updated Script\UI\InspectPage.ts;
Updated the label 'Profit' element on InspectPage for buildings so that it is now labelled 'Profit (Auto Sell)'. This change was made to better make clear that the calculation is based on local market prices not the player market.
Thanks to existing localization entries; this modification was accomplished via the following operations t("BuildingProfit")+" ("+t("AutoSell")+")
As a result no localization updates are required.